.NET: chunked buffer + int64 offsets, removing the 2GB model.dat ceiling
.NET's MemoryStream/byte[] has a hard ~2.1GB single-array limit that no GC
setting can lift - real production files with model.dat decompressing
past that (observed at 3-6GB on the largest test files, ~10x compression
ratio) hit "Stream was too long" / OutOfMemoryException within seconds,
regardless of available system RAM.

Added ChunkedBuffer: a byte buffer backed by multiple bounded byte[]
segments instead of one contiguous array, sized from the ZIP entry's own
recorded uncompressed length so each segment allocates once at its final
size. Widened TlvNode.Offset/Size and the TLV parsing functions from int
to long to address into it. Geometry extraction logic is untouched - node
payloads are still small, ordinary byte[] slices.

Verified: full test suite still green (4/4), zero regressions across all
previously-working files (exact same counts as before), and all three
files that used to crash outright (294MB, 392MB, 620MB) now parse
successfully - including the 620MB file, which no other language in this
project currently completes. Cross-checked against Dart's independently
-computed results for the same two mid files: identical down to every
field.

+++ packages/dotnet/OpenSkp/Geometry.cs
@@ -19,7 +19,7 @@ namespace OpenSkp
 
     internal sealed class GeometryBuilderInstance
     {
-        public int Offset;
+        public long Offset;
         public string? RefGuid;
         public long? RefIdx;
         public string? Name;
